Assignment 3 ready for demo

+++ Assignment3/Spring-2025-Demo-3-REST-master/src/main/java/org/example/spring2025demo3rest/dataaccess/HomeRepository.java
@@ -8,21 +8,21 @@ import java.util.Optional;
 /**
  * Home Repository interface that will be used by String to create a bean that handles all the CRUD operations
  */
-public interface HomeRepository extends CrudRepository<Home, Long> {
+public interface HomeRepository extends CrudRepository<Home, Integer> {
 
     /**
      * Get all homes for a user
      * @param userId
      * @return
      */
-    Iterable<Home> getAllByUserId(Long userId);
+    Iterable<Home> getAllByUserId(Integer userId);
 
     /**
      * Get user by ID
      * @param userId
      * @return
      */
-    Optional<Home> findByUserId(Long userId);
+    Optional<Home> findByUserId(Integer userId);
 
 
     //Note: I am surprised this works! Spring detects it and fills in the method. Crazy.
