+++ Assignment3/Spring-2025-Demo-3-REST-master/src/main/java/org/example/spring2025demo3rest/dataaccess/UserRepository.java
@@ -6,9 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 /**
  * User Repository interface that will be used by String to create a bean that handles all the CRUD operations
  */
-public interface UserRepository extends CrudRepository<User, Long> {
+public interface UserRepository extends CrudRepository<User, Integer> {
     User getUserById(Integer id);
-
-    //This is where you write code needed beyond the basics
 
 }
